refactoring and backend
refactored some file structure and added new files which will be used to implement backend authentication adn user signin and registration.
