+++ flutter_app/lib/authentication.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class Authentication extends StatefulWidget {
+  @override
+  _AuthenticationState createState() => _AuthenticationState();
+}
+
+class _AuthenticationState extends State<Authentication> {
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      child: Text('Authenticate'),
+    );
+  }
+}
