+++ flutter_app/lib/traffic.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_app/authentication.dart';
+import 'package:flutter_app/homePage.dart';
+
+class Traffic extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    //return home or authentication screen
+    return LandingPage();
+  }
+}
