+++ flutter_app/pubspec.yaml
@@ -21,7 +21,7 @@ environment:
   sdk: ">=2.7.0 <3.0.0"
 
 dependencies:
-  flutter:
+  flutter:  
     sdk: flutter
   app_settings: 4.0.4
   app_usage: 1.2.0
